requirements-all.txt
a2wsgi==1.10.8
aiobotocore==2.12.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
bentoml==1.4.12
billiard==4.2.1
boto3==1.34.51
botocore==1.34.51
cattrs==23.1.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.7
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
common_code @ git+https://github.com/swiss-ai-center/common-code.git@6445b2305a299afbdbdd70c9efd7c94865cc91d6
configobj==5.0.9
contourpy==1.3.2
coverage==7.8.0
cryptography==44.0.3
cycler==0.12.1
Deprecated==1.2.18
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dpath==2.2.0
dulwich==0.22.8
dvc==3.58.0
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-s3==3.2.0
dvc-studio-client==0.21.0
dvc-task==0.40.2
email_validator==2.1.1
entrypoints==0.4
fastapi==0.110.0
filelock==3.18.0
flake8==7.2.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.2
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
grandalf==0.8
gto==1.7.2
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
httpx-ws==0.7.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kantoku==0.18.3
kiwisolver==1.4.8
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-aiohttp-client==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pip==24.2
pip-requirements-parser==32.0.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.2.1
pydantic_core==2.33.2
pydot==4.0.0
pyflakes==3.3.2
pygit2==1.18.0
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.3
pypdfium2==4.30.1
pytest==8.1.1
pytest-asyncio==0.23.5.post1
pytest-cov==4.1.0
pytest_httpserver==1.0.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.4.0
questionary==2.1.0
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3fs==2025.3.2
s3transfer==0.10.4
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.3
scmrepo==3.3.11
semver==3.0.4
setuptools==80.7.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simple-di==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
starlette==0.36.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tomli_w==1.2.0
tomlkit==0.13.2
torch==2.2.2
torchvision==0.17.2
tornado==6.5
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.0.7
uvicorn==0.28.0
vine==5.1.0
voluptuous==0.15.2
watchfiles==1.0.5
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zc.lockfile==3.0.post1
zipp==3.21.0
